🗃 update and change location of entitites
